Extract shared uploader modules from the two entry points
RuntimeUploaderConcern (controllers) and UploaderHelper (views, ActiveJobs,
standalone objects) carried 349 byte-identical lines across 18 methods, so
every upload fix had to be written twice and Metrics/ModuleLength had been
raised three times to accommodate them.

Move the shared bodies into UploaderPipeline, UploaderImageProcessing, and
UploaderSupport under lib/camaleon_cms/, which both entry points include --
the same pattern as UploaderPathSecurity and UploaderContentSecurity. The
split is three ways rather than one because a single 349-line module would
itself be the longest in the repo, leaving the ModuleLength ceiling where
it was.

Both entry points keep their full public API: every method retains its name,
arity, visibility, and return value. Deleting either one is not an option --
custom_initializers.rb documents host apps including UploaderHelper inside an
ActiveJob, which helper_method exposure would not reach.

Three behavior-relevant details:

Messages. User-facing upload errors differ by context, so the pipeline renders
them through cama_uploader_ct / cama_uploader_t / cama_uploader_human_size.
The defaults go to I18n, which is all a controller can do (CamaleonController
does not include CamaleonHelper, so `ct` is undefined there); UploaderHelper
overrides all three to keep routing through ct/cama_t, whose on_translation
hook lets plugins override the text.

Staging name. cama_stage_data_uri guarded on params[:name], then immediately
used args[:name]. It now reads args[:name]. This is a no-op on every
controller path, since MediaController passes name: params[:name] through.
Off that path it makes the helper's data: branch work in an ActiveJob, where
params is undefined and the call previously raised NameError despite
custom_initializers.rb documenting that usage. Keeping params[:name] would
enshrine a request-context dependency in shared code purely to preserve a
latent bug.

Settings. Both entry points now normalize with to_h.deep_symbolize_keys before
assigning :uploaded_io. This keeps the helper's existing depth
(deep_symbolize_keys and the lib/ext Hash#to_sym monkeypatch produce ==
results, and the former does not raise on a non-symbolizable key) and drops
the extra caller-hash mutation the concern performed.

Metrics/ModuleLength drops 364 -> 326, now set by admin/custom_fields_helper.rb
with no uploader involvement, and the Naming/MethodParameterName exclusion for
uploader_helper.rb is no longer needed.
