+++ app/controllers/concerns/camaleon_cms/runtime_uploader_concern.rb
@@ -1,331 +1,16 @@
-require 'net/http'
-require 'tempfile'
-
-# rubocop:disable Naming/MethodParameterName
 module CamaleonCms
+  # Uploader entry point for controllers (mixed into CamaleonCms::CamaleonController and
+  # re-exported through RuntimeStateConcern). The implementation lives in the shared
+  # modules below, which CamaleonCms::UploaderHelper includes as well, so the two entry
+  # points cannot drift. Only cama_upload_url_error is controller-specific.
   module RuntimeUploaderConcern
     extend ActiveSupport::Concern
 
     include UploaderContentSecurity
     include UploaderPathSecurity
-
-    def upload_file(uploaded_io, settings = {})
-      cached_name = uploaded_io.is_a?(ActionDispatch::Http::UploadedFile) ? uploaded_io.original_filename : nil
-      return { error: 'File is empty', file: nil, size: nil } if uploaded_io.blank?
-
-      if uploaded_io.is_a?(String) && uploaded_io.match(%r{^https?://}).present?
-        tmp = cama_tmp_upload(uploaded_io)
-        return tmp if tmp[:error].present?
-
-        settings[:remove_source] = true
-        uploaded_io = tmp[:file_path]
-      end
-      if uploaded_io.is_a?(String)
-        expanded = cama_canonical_upload_path(uploaded_io)
-        return { error: 'Invalid file path' } unless expanded
-
-        uploaded_io = expanded
-      end
-      uploaded_io = File.open(uploaded_io) if uploaded_io.is_a?(String)
-      if settings[:dimension].present?
-        uploaded_io = File.open(cama_resize_upload(uploaded_io.path, settings[:dimension]))
-      end
-
-      if file_content_unsafe?(uploaded_io)
-        return cama_upload_failure({ error: 'Potentially malicious content found!' }, uploaded_io, settings)
-      end
-
-      settings[:uploaded_io] = uploaded_io
-      settings = settings.to_h.symbolize_keys
-      settings = {
-        folder: '',
-        maximum: current_site.get_option('filesystem_max_size', 100).to_f.megabytes,
-        formats: '*',
-        generate_thumb: true,
-        temporal_time: 0,
-        filename: begin
-          cached_name || uploaded_io.original_filename
-        rescue StandardError
-          uploaded_io.path.split('/').last
-        end.cama_fix_filename,
-        file_size: File.size(uploaded_io.to_io),
-        remove_source: false,
-        same_name: false,
-        versions: '',
-        thumb_size: nil
-      }.merge!(settings)
-      settings[:formats] = '*' if settings[:formats].nil?
-      settings[:folder] = '' if settings[:folder].nil? # e.g. crop_url passes no folder
-      hooks_run('before_upload', settings)
-
-      unless cama_uploader.valid_folder_path?(settings[:folder])
-        return cama_upload_failure({ error: 'Invalid file path' }, uploaded_io, settings)
-      end
-
-      err = validate_file_format_or_error(uploaded_io.path, settings[:formats])
-      return cama_upload_failure(err, uploaded_io, settings) if err
-
-      err = cama_size_limit_error(settings[:file_size], settings[:maximum])
-      return cama_upload_failure(err, uploaded_io, settings) if err
-
-      key = File.join(settings[:folder], settings[:filename]).to_s.cama_fix_slash
-      res = cama_uploader.add_file(settings[:uploaded_io], key, { same_name: settings[:same_name] })
-
-      if res['file_type'] == 'image'
-        settings[:versions].to_s.delete(' ').split(',').each do |v|
-          version_path = cama_resize_upload(settings[:uploaded_io].path, v, { replace: false })
-          cama_uploader.add_file(version_path, cama_uploader.version_path(res['key'], v), is_thumb: true,
-                                                                                          same_name: true)
-          FileUtils.rm_f(version_path)
-        end
-      end
-
-      if settings[:generate_thumb] && res['thumb'].present?
-        cama_uploader_generate_thumbnail(uploaded_io.path, res['key'], settings[:thumb_size],
-                                         settings[:remove_source])
-      end
-      FileUtils.rm_f(uploaded_io.path) if settings[:remove_source] && File.exist?(uploaded_io.path)
-
-      hooks_run('after_upload', settings)
-      CamaleonCmsUploader.delete_block.call(settings, cama_uploader, key) if settings[:temporal_time] > 0
-
-      res
-    end
-
-    def cama_uploader_generate_thumbnail(uploaded_io, key, thumb_size = nil, remove_source = false)
-      w = thumb_size.present? ? thumb_size.split('x')[0] : cama_uploader.thumb[:w]
-      h = thumb_size.present? ? thumb_size.split('x')[1] : cama_uploader.thumb[:h]
-      uploaded_io = File.open(uploaded_io) if uploaded_io.is_a?(String)
-      path_thumb = cama_resize_and_crop(uploaded_io.path, w, h)
-      thumb = cama_uploader.add_file(path_thumb, cama_uploader.version_path(key).sub('.svg', '.jpg'), is_thumb: true,
-                                                                                                      same_name: true)
-      FileUtils.rm_f(path_thumb) if remove_source
-      thumb
-    end
-
-    def uploader_verify_name(file_path)
-      dir = File.dirname(file_path)
-      filename = File.basename(file_path).to_s.cama_fix_filename
-      files = Dir.entries(dir)
-      if files.include?(filename)
-        i = 1
-        _filename = filename
-        while files.include?(_filename)
-          _filename = "#{File.basename(filename, File.extname(filename))}_#{i}#{File.extname(filename)}"
-          i += 1
-        end
-        filename = _filename
-      end
-      "#{File.dirname(file_path)}/#{filename}"
-    end
-
-    def cama_file_path_to_url(file_path)
-      file_path.sub(Rails.public_path.to_s, begin
-        root_url
-      rescue StandardError
-        cama_root_url
-      end)
-    end
-
-    def cama_url_to_file_path(url)
-      File.join(Rails.public_path, URI(url.to_s).path)
-    end
-
-    def cama_crop_image(file_path, w = nil, h = nil, w_offset = 0, h_offset = 0, resize = false, replace = true)
-      force = w.present? && h.present? && !w.include?('?') && !h.include?('?') ? '!' : ''
-      img = MiniMagick::Image.open(file_path)
-      w = clamp_to_image_dimension(w, img[:width])
-      h = clamp_to_image_dimension(h, img[:height])
-      data = { img: img, w: w, h: h, w_offset: w_offset, h_offset: h_offset, resize: resize, replace: replace }
-      hooks_run('before_crop_image', data)
-      data[:img].combine_options do |i|
-        i.resize("#{w.presence}x#{h.presence}#{force}") if data[:resize]
-        i.crop "#{w.presence}x#{h.presence}+#{w_offset}+#{h_offset}#{force}" unless data[:resize]
-      end
-
-      ext = File.extname(file_path)
-      res = data[:replace] ? file_path : file_path.gsub(ext, "_crop#{ext}")
-      data[:img].write res
-      res
-    end
-
-    def cama_resize_and_crop(file, w, h, settings = {})
-      settings = { gravity: :north_east, overwrite: true, output_name: +'' }.merge!(settings)
-      img = MiniMagick::Image.open(file)
-      if file.end_with? '.svg'
-        img.format 'jpg'
-        file.sub! '.svg', '.jpg'
-        settings[:output_name]&.sub!('.svg', '.jpg')
-      end
-      w = clamp_to_image_dimension(w, img[:width])
-      h = clamp_to_image_dimension(h, img[:height])
-      w_original = img[:width].to_f
-      h_original = img[:height].to_f
-      w = w.to_i if w.present?
-      h = h.to_i if h.present?
-
-      if w_original * h < h_original * w
-        op_resize = "#{w.to_i}x"
-        w_result = w
-        h_result = (h_original * w / w_original)
-      else
-        op_resize = "x#{h.to_i}"
-        w_result = (w_original * h / h_original)
-        h_result = h
-      end
-
-      w_offset, h_offset = cama_crop_offsets_by_gravity(settings[:gravity], [w_result, h_result], [w, h])
-      data = { img: img, w: w, h: h, w_offset: w_offset, h_offset: h_offset, op_resize: op_resize, settings: settings }
-      hooks_run('before_resize_crop', data)
-      data[:img].combine_options do |i|
-        i.resize(data[:op_resize])
-        i.gravity(settings[:gravity])
-        i.crop "#{data[:w].to_i}x#{data[:h].to_i}+#{data[:w_offset]}+#{data[:h_offset]}!"
-      end
-
-      if settings[:overwrite]
-        data[:img].write(file.sub('.svg', '.jpg'))
-      elsif settings[:output_name].present?
-        data[:img].write(file = File.join(File.dirname(file), settings[:output_name]).to_s)
-      else
-        data[:img].write(file = uploader_verify_name(File.join(File.dirname(file),
-                                                               "crop_#{File.basename(file.sub('.svg', '.jpg'))}")))
-      end
-      file
-    end
-
-    def cama_tmp_upload(uploaded_io, args = {})
-      tmp_path = args[:path] || File.join(Rails.public_path, 'tmp', current_site.id.to_s).to_s
-      FileUtils.mkdir_p(tmp_path)
-      # Default to the site limit so the size guard below actually applies: callers
-      # such as crop/crop_url pass no :maximum, which left it dead code.
-      args[:maximum] ||= current_site.get_option('filesystem_max_size', 100).to_f.megabytes
-      saved = false
-      downloaded_tmp_file = nil
-      staged_path = nil
-      if uploaded_io.is_a?(String) && uploaded_io.start_with?('data:')
-        path, err = cama_stage_data_uri(uploaded_io, args, tmp_path)
-        return err if err
-
-        staged_path = path
-        _tmp_name = File.basename(args[:name].to_s)
-        uploaded_io = File.open(path)
-        saved = true
-      elsif uploaded_io.is_a?(String) && uploaded_io.start_with?('http://', 'https://')
-        err = validate_file_format_or_error(uploaded_io, args[:formats])
-        return err if err
-
-        if same_site_url?(uploaded_io, current_site)
-          uploaded_io = File.join(Rails.public_path, site_url_path(uploaded_io, current_site)).to_s
-        else
-          remote_file = cama_download_remote_file(uploaded_io)
-          return remote_file if remote_file[:error].present?
-
-          downloaded_tmp_file = remote_file[:file]
-          uploaded_io = downloaded_tmp_file
-        end
-        _tmp_name = if uploaded_io.is_a?(String)
-                      uploaded_io.split('/').last.split('?').first
-                    else
-                      uploaded_io.path.split('/').last
-                    end
-        args[:name] = args[:name] || _tmp_name
-      end
-      if uploaded_io.is_a?(String)
-        expanded = cama_canonical_upload_path(uploaded_io)
-        return { error: 'Invalid file path' } unless expanded
-
-        uploaded_io = expanded
-      end
-      uploaded_io = File.open(uploaded_io) if uploaded_io.is_a?(String)
-      err = validate_file_format_or_error(_tmp_name || uploaded_io.path, args[:formats])
-      return err if err
-
-      actual_size = begin
-        uploaded_io.size
-      rescue StandardError
-        File.size(uploaded_io)
-      end
-      err = cama_size_limit_error(actual_size, args[:maximum])
-      return err if err
-
-      name = args[:name] || uploaded_io&.original_filename || uploaded_io.path.split('/').last
-      name = "#{File.basename(name, File.extname(name)).parameterize}#{File.extname(name)}"
-      path ||= uploader_verify_name(File.join(tmp_path, name))
-      unless saved
-        # Same rule as the data: branch above -- read, scan, and only then write, so
-        # a remote or same-site source cannot land unscanned in public/tmp.
-        content = uploaded_io.read
-        return { error: 'Potentially malicious content found!' } if content_unsafe?(content, filename: name)
-
-        File.open(path, 'wb') { |f| f.write(content) }
-        staged_path = path
-      end
-      path = cama_resize_upload(path, args[:dimension]) if args[:dimension].present?
-      { file_path: path, error: nil }
-    rescue StandardError
-      # A raised error leaves no half-written file behind in the served staging dir.
-      cama_purge_staged_file(staged_path, tmp_path)
-      raise
-    ensure
-      downloaded_tmp_file&.close!
-    end
-
-    def cama_resize_upload(image_path, dimension, args = {})
-      if cama_uploader.class.validate_file_format(image_path, 'image') && dimension.present?
-        dim_parts = dimension.split('x')
-        r = { file: image_path, w: dim_parts[0], h: dim_parts[1], w_offset: 0, h_offset: 0,
-              resize: !dim_parts[2] || dim_parts[2] == 'resize',
-              replace: true, gravity: :north_east }.merge!(args)
-        hooks_run('on_uploader_resize', r)
-        image_path = if r[:w].present? && r[:h].present?
-                       cama_resize_and_crop(r[:file], r[:w], r[:h], { overwrite: r[:replace], gravity: r[:gravity] })
-                     else
-                       cama_crop_image(r[:file], r[:w], r[:h], r[:w_offset], r[:h_offset], r[:resize], r[:replace])
-                     end
-      end
-      image_path
-    end
-
-    def cama_uploader
-      @cama_uploader ||= lambda {
-        thumb = current_site.get_option('filesystem_thumb_size', '100x100').split('x')
-        args = {
-          server: current_site.get_option('filesystem_type', 'local').downcase,
-          thumb: { w: thumb[0], h: thumb[1] },
-          aws_settings: {
-            region: current_site.get_option('filesystem_region', 'us-west-2'),
-            access_key: current_site.get_option('filesystem_s3_access_key'),
-            secret_key: current_site.get_option('filesystem_s3_secret_key'),
-            bucket: current_site.get_option('filesystem_s3_bucket_name'),
-            cloud_front: current_site.get_option('filesystem_s3_cloudfront'),
-            aws_file_upload_settings: ->(settings) { settings },
-            aws_file_read_settings: ->(data, _s3_file) { data }
-          },
-          custom_uploader: nil
-        }
-        hooks_run('on_uploader', args)
-        return args[:custom_uploader] if args[:custom_uploader].present?
-
-        base_args = { current_site: current_site, thumb: args[:thumb] }
-        case args[:server]
-        when 's3', 'aws'
-          CamaleonCmsAwsUploader.new(base_args.merge(aws_settings: args[:aws_settings]), self)
-        else
-          CamaleonCmsLocalUploader.new(base_args, self)
-        end
-      }.call
-    end
-
-    def slugify(val)
-      val.to_s.downcase.strip.tr(' ', '-').gsub(/[^\w-]/, '')
-    end
-
-    def slugify_folder(val)
-      split_folder = val.split('/')
-      split_folder[-1] = slugify(split_folder[-1])
-      split_folder.join('/')
-    end
+    include UploaderPipeline
+    include UploaderImageProcessing
+    include UploaderSupport
 
     private
 
@@ -353,115 +38,5 @@ module CamaleonCms
                end
       result.join(', ') if result.is_a?(Array)
     end
-
-    def cama_download_remote_file(url)
-      validator = UserUrlValidator.new
-      validation_result = validator.validate(url, reject_path_traversal: true)
-      return { error: validation_result.join(', ') } if validation_result.is_a?(Array)
-
-      uri = URI.parse(url)
-      http = Net::HTTP.new(uri.host, uri.port)
-      http.use_ssl = uri.scheme == 'https'
-      # Pin the socket to the exact IP the validator vetted so a hostname whose DNS
-      # answer changes between the SSRF check and the fetch (DNS rebinding) cannot
-      # redirect the request to an internal address; SNI/cert checks keep the host.
-      http.ipaddr = validator.resolved_ip if validator.resolved_ip
-      http.open_timeout = 10
-      http.read_timeout = 10
-      response = http.start do |conn|
-        conn.request(Net::HTTP::Get.new(uri.request_uri.presence || '/'))
-      end
-
-      return { error: 'Redirects are not allowed for remote uploads.' } if response.is_a?(Net::HTTPRedirection)
-      unless response.is_a?(Net::HTTPSuccess)
-        return { error: "Unable to download remote file (HTTP #{response.code})." }
-      end
-
-      max_bytes = current_site.get_option('filesystem_max_size', 100).to_f.megabytes
-      body = response.body
-      if body.bytesize > max_bytes
-        return { error: "Remote file too large (max #{ActiveSupport::NumberHelper.number_to_human_size(max_bytes)})." }
-      end
-
-      ext = File.extname(uri.path.to_s)
-      tempfile = Tempfile.new(['cama-upload-url', ext], binmode: true)
-      tempfile.write(body)
-      tempfile.rewind
-      { file: tempfile, error: nil }
-    rescue StandardError => e
-      { error: "Unable to download remote file: #{ERB::Util.html_escape(e.message)}" }
-    end
-
-    def cama_crop_offsets_by_gravity(gravity, original_dimensions, cropped_dimensions)
-      original_width, original_height = original_dimensions
-      cropped_width, cropped_height = cropped_dimensions
-
-      vertical_offset = case gravity
-                        when :north_west, :north, :north_east then 0
-                        when :center, :east, :west then [((original_height - cropped_height) / 2.0).to_i, 0].max
-                        when :south_west, :south, :south_east then (original_height - cropped_height).to_i
-                        end
-
-      horizontal_offset = case gravity
-                          when :north_west, :west, :south_west then 0
-                          when :center, :north, :south then [((original_width - cropped_width) / 2.0).to_i, 0].max
-                          when :north_east, :east, :south_east then (original_width - cropped_width).to_i
-                          end
-
-      [horizontal_offset, vertical_offset]
-    end
-
-    def clamp_to_image_dimension(value, img_size)
-      return value unless value.present? && value.to_s.include?('?')
-
-      img_size.to_f > value.sub('?', '').to_i ? value.sub('?', '') : img_size
-    end
-
-    def validate_file_format_or_error(file, formats)
-      return if cama_uploader.class.validate_file_format(file, formats)
-
-      { error: "#{I18n.t('camaleon_cms.common.file_format_error')} (#{formats})" }
-    end
-
-    # Stages a base64 data: payload into tmp_path, bounding its size and scanning its
-    # content BEFORE the write, so nothing oversized or hostile ever reaches a path
-    # the web server would hand out. Returns [path, error]; exactly one is non-nil.
-    def cama_stage_data_uri(uploaded_io, args, tmp_path)
-      return [nil, { error: I18n.t('camaleon_cms.admin.media.name_required').to_s }] if params[:name].blank?
-
-      # Strip any directory components so a hostile name (e.g. "../../etc/x")
-      # cannot escape tmp_path when the base64 payload is written below.
-      tmp_name = File.basename(args[:name].to_s)
-      err = validate_file_format_or_error(tmp_name, args[:formats])
-      return [nil, err] if err
-
-      payload = uploaded_io.split(';base64,').last
-      # Bound the payload before decoding, so an oversized upload is never allocated
-      # in full nor written into the served staging directory.
-      err = cama_size_limit_error(cama_base64_decoded_size(payload), args[:maximum])
-      return [nil, err] if err
-
-      # The decoded bytes are already in memory here, so scanning costs no extra copy.
-      decoded = Base64.decode64(payload)
-      return [nil, { error: 'Potentially malicious content found!' }] if content_unsafe?(decoded, filename: tmp_name)
-
-      path = uploader_verify_name(File.join(tmp_path, tmp_name))
-      return [nil, { error: 'Invalid file path' }] unless path_within?(path, tmp_path)
-
-      File.open(path, 'wb') { |f| f.write(decoded) }
-      [path, nil]
-    end
-
-    # Returns an error hash when size exceeds maximum, nil otherwise. Not shared with
-    # UploaderHelper's copy: that one translates via `ct`, which runs the
-    # `on_translation` hook so plugins can override the message. Unifying them would
-    # silently drop that hook on one side.
-    def cama_size_limit_error(size, maximum)
-      return if maximum.blank? || maximum >= size
-
-      max_size = ActiveSupport::NumberHelper.number_to_human_size(maximum)
-      { error: "#{I18n.t('camaleon_cms.common.file_size_exceeded', default: 'File size exceeded')} (#{max_size})" }
-    end
   end
 end
-# rubocop:enable Naming/MethodParameterName
