+++ lib/camaleon_cms.rb
@@ -5,6 +5,9 @@ require 'camaleon_cms/svg_content_checker'
 require 'camaleon_cms/media_security_headers'
 require 'camaleon_cms/uploader_content_security'
 require 'camaleon_cms/uploader_path_security'
+require 'camaleon_cms/uploader_pipeline'
+require 'camaleon_cms/uploader_image_processing'
+require 'camaleon_cms/uploader_support'
 
 module CamaleonCms
 end
